move content store fitures and helpers to a separate file
